Allow return of function call to be returned by generateText
Hi !

generateText method does currently not return the functions calls
results.

When calling generateText, the LLM does not return anything but an empty
string. The generateText caller have no clue on what was done.

This PR propose to get results from one or more function calls (with
handleTools) and convert it in JSON when the LLM result is empty.

I think it can be used by the caller, in a second time, to provide more
context to the LLM on a later call.

Tell me what you think, tanks !